fix(3p/cppnix): ignore the enforceDeterminism value
We used to set enforceDeterminism to true in the settings (by default)
and thus did send a non-zero value over the wire. The value should
probably be ignored as it should only matter if nrRounds is non-zero
as well.

Having the old code here where the value is expected to be zero only
works with the same version of Nix where we are sending zero. We
should always test this against older Nix versions being client or
server as otherwise upgrade in larger networks might be a pain.

See https://github.com/NixOS/nix/pull/7099#issuecomment-1344943739

Change-Id: I1353cc9eff6add4a7762aa41e6b8efe78c3e04ac
Reviewed-on: https://cl.tvl.fyi/c/depot/+/14114
Tested-by: BuildkiteCI
Reviewed-by: tazjin <tazjin@tvl.su>
Autosubmit: sterni <sternenseemann@systemli.org>