fix(3p/cppnix): MonitorFdHup: Make it work on macOS again
It appears that on current macOS versions, our use of poll() to detect
client disconnects no longer works. As a workaround, poll() for
POLLRDNORM, since this *will* wake up when the client has
disconnected. The downside is that it also wakes up when input is
available. So just sleep for a bit in that case.  This means that on
macOS, a client disconnect may take up to a second to be detected,
but that's better than not being detected at all.

Fixes #7584.

(cherry picked from commit 9fc8d00d741d17ff25c0193f182103d00d11582a)

Nixpkgs applied this as a patch on top of nix_2_3, so let's pick this
onto our branch, so we don't see regressions for Darwin users in depot
once we have to package it ourselves again.

Change-Id: Ie24e73d7e717dacdf6ee2e18fcc799ec4c366855
Reviewed-on: https://cl.tvl.fyi/c/depot/+/13857
Autosubmit: sterni <sternenseemann@systemli.org>
Tested-by: BuildkiteCI
Reviewed-by: tazjin <tazjin@tvl.su>