fix(3p/cppnix): set runtime relevant directories explicitly
I elided these configureFlags in cl/13652, assuming the arguments
existed mostly for the sake of configurability. However, it turns out
that the Nix configure script may pick bad defaults (based on --prefix)
if this is not given explicitly.

In particular, it would default to $out/etc for sysconfdir and
completely ignore /etc/nix/nix.conf.

Resolves b/454.

Change-Id: I4ea7279b4760fcca499b334cbcdc657e99f7fb8d
Reviewed-on: https://cl.tvl.fyi/c/depot/+/13731
Reviewed-by: tazjin <tazjin@tvl.su>
Autosubmit: sterni <sternenseemann@systemli.org>
Tested-by: BuildkiteCI